app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
duckdb==0.10.2
pandas==2.2.2
orjson==3.10.3
python-multipart==0.0.9
openai==1.30.1
python-dotenv==1.0.1
